feat: better version api
