+++ src/v2/interfaces/versions.ts
@@ -1,0 +1,7 @@
+import { Edition } from "./textures";
+
+export interface NewVersionParam {
+	edition: Edition;
+	version: string;
+	newVersion: string;
+}
