+++ src/v2/service/path.service.ts
@@ -1,11 +1,8 @@
 import { WriteConfirmation } from "firestorm-db";
-import { BadRequestError, NotFoundError } from "../tools/errorTypes";
+import { BadRequestError } from "../tools/errorTypes";
 import UseService from "./use.service";
-import { InputPath, Path, PathNewVersionParam } from "../interfaces";
+import { InputPath, Path } from "../interfaces";
 import PathFirestormRepository from "../repository/path.repository";
-import TextureService from "./texture.service";
-import { settings } from "../firestorm";
-import versionSorter from "../tools/versionSorter";
 
 export default class PathService {
 	private readonly useService: UseService;
@@ -58,67 +55,6 @@ export default class PathService {
 
 		await this.useService.getUseByIdOrName(path.use); // verify use existence
 		return this.repo.updatePath(id, path);
-	}
-
-	async addVersion(body: PathNewVersionParam): Promise<[WriteConfirmation, WriteConfirmation]> {
-		// stupid workaround for recursion (the classes require each other)
-		const versions = await TextureService.getInstance().getVersionByEdition(body.edition);
-
-		// check existing version to the paths provided
-		if (!versions.includes(body.version))
-			throw new BadRequestError("Incorrect input path version provided");
-
-		return Promise.all([
-			settings.editField({
-				id: "versions",
-				field: body.edition,
-				operation: "array-splice",
-				// equivalent of array_unshift (new versions go at start of list)
-				value: [0, 0, body.newVersion],
-			}),
-			this.repo.addNewVersionToVersion(body.version, body.newVersion),
-		]);
-	}
-
-	async removeVersion(version: string): Promise<[WriteConfirmation, WriteConfirmation]> {
-		const allVersions: Record<string, string[]> = await settings.get("versions");
-		const edition = Object.entries(allVersions).find((v) => v[1].includes(version))?.[0];
-
-		if (!edition) throw new NotFoundError(`Matching edition not found for version ${version}`);
-
-		return Promise.all([
-			settings.editField({
-				id: "versions",
-				field: edition,
-				operation: "set",
-				value: allVersions[edition].filter((v) => v !== version),
-			}),
-			this.repo.removeVersion(version),
-		]);
-	}
-
-	async renameVersion(
-		oldVersion: string,
-		newVersion: string,
-	): Promise<[WriteConfirmation, WriteConfirmation]> {
-		const allVersions: Record<string, string[]> = await settings.get("versions");
-		const edition = Object.entries(allVersions).find((v) => v[1].includes(oldVersion))?.[0];
-		if (!edition) throw new NotFoundError(`Matching edition not found for version ${oldVersion}`);
-
-		return Promise.all([
-			settings.editField({
-				id: "versions",
-				field: edition,
-				operation: "set",
-				// map old version to new version, keep the rest the same
-				value: allVersions[edition]
-					.map((v) => (v === oldVersion ? newVersion : v))
-					.sort(versionSorter)
-					// newest at top
-					.reverse(),
-			}),
-			this.repo.renameVersion(oldVersion, newVersion),
-		]);
 	}
 
 	removePathById(pathID: string): Promise<WriteConfirmation> {
