+++ tsconfig.json
@@ -16,6 +16,7 @@
 		"declaration": true,
 		"allowJs": true,
 		"strictNullChecks": true,
+		"strict": false,
 		"resolveJsonModule": true,
 	},
 	"exclude": [
